Moved code around to provide clarity around the kind of traffic being displayed
