+++ library/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             android:launchMode="singleTask"
             android:taskAffinity="com.chuckerteam.chucker.task"
             android:theme="@style/Chucker.Theme" />
-        <activity android:name="com.chuckerteam.chucker.internal.ui.transaction.TransactionActivity"
+        <activity android:name="com.chuckerteam.chucker.internal.ui.traffic.http.TransactionActivity"
             android:theme="@style/Chucker.Theme"
             android:parentActivityName="com.chuckerteam.chucker.internal.ui.MainActivity"/>
         <activity android:name="com.chuckerteam.chucker.internal.ui.error.ErrorActivity"
