+++ library/src/main/java/com/chuckerteam/chucker/internal/ui/traffic/http/TransactionOverviewFragment.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.chuckerteam.chucker.internal.ui.transaction;
+package com.chuckerteam.chucker.internal.ui.traffic.http;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
