+++ library/src/main/java/com/chuckerteam/chucker/internal/ui/traffic/websocket/WebsocketTrafficViewHolder.kt
@@ -1,4 +1,4 @@
-package com.chuckerteam.chucker.internal.ui.transaction
+package com.chuckerteam.chucker.internal.ui.traffic.websocket
 
 import android.view.View
 import android.widget.TextView
@@ -7,6 +7,8 @@ import com.chuckerteam.chucker.internal.data.entity.TrafficType
 import com.chuckerteam.chucker.internal.data.entity.TrafficType.WEBSOCKET_TRAFFIC
 import com.chuckerteam.chucker.internal.data.entity.WebsocketTraffic
 import com.chuckerteam.chucker.internal.support.formatBytes
+import com.chuckerteam.chucker.internal.ui.traffic.TrafficRow
+import com.chuckerteam.chucker.internal.ui.traffic.TrafficViewHolder
 import java.text.DateFormat
 
 class WebsocketTrafficViewHolder(view: View, listener: (Long, Int, TrafficType) -> Unit) :
@@ -31,7 +33,8 @@ class WebsocketTrafficViewHolder(view: View, listener: (Long, Int, TrafficType) 
 }
 
 @Suppress("EqualsOrHashCode")
-internal class WebsocketTrafficRow(val traffic: WebsocketTraffic) : TrafficRow {
+internal class WebsocketTrafficRow(val traffic: WebsocketTraffic) :
+    TrafficRow {
     override val id: Long = traffic.id
     override val timestamp: Long = traffic.timestamp ?: 0L
     override val type = WEBSOCKET_TRAFFIC
