+++ library/src/main/res/layout/chucker_activity_transaction.xml
@@ -21,7 +21,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layoutDirection="ltr"
-    tools:context="com.chuckerteam.chucker.internal.ui.transaction.TransactionActivity">
+    tools:context="com.chuckerteam.chucker.internal.ui.traffic.http.TransactionActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appbar"
